+++ src/main/java/com/myParty/controllers/MembersController.java
@@ -165,29 +165,7 @@ public class MembersController {
         return "redirect:/";
     }
 
-    @GetMapping("/member/resetpassword/{token}")
-    //TODO: make sure this works and create the form
-    public String showResetPWform(@PathVariable("token") String token) {
-        Member member = memberDao.findByResetToken(token);
-        return "member/resetpassword";
-    }
 
-    @PostMapping("/member/resetpassword")
-    public String resetPassword(HttpServletRequest request,
-                                         @RequestParam("email") String memberEmail) throws MessagingException {
-        Member member = memberDao.getByEmail(memberEmail);
-//        if (member == null) {
-//            throw new UserNotFoundException();
-//        }
-        String token = UUID.randomUUID().toString();
-        member.setResetToken(token);
-        memberDao.save(member);
-        String resetDetails = "Click here to reset your password";
-//TODO: add domain + the token in an anchor tag
-        emailService.sendResetPassword(member, resetDetails);
-
-        return "";
-    }
 }
