+++ src/main/resources/templates/member/resetpassword.html
@@ -1,10 +1,21 @@
 <!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<!--RESET PASSWORD FORM GOES HERE-->
-</body>
+<html xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout}">
+
+<meta charset="UTF-8"/>
+<title layout:fragment="title">myParty | Reset Password</title>
+
+<section layout:fragment="content">
+
+    <h1>Reset Your Password</h1>
+
+<!--    TODO: change th:action-->
+    <form th:action="@{/sign-up}" method="POST" th:object="${member}">
+
+        <label for="password">Password</label>
+        <input id="password" type="password" th:field="*{password}"/>
+
+        <input value="Reset" type="submit"/>
+    </form>
+</section>
+<footer><p layout:fragment="custom-footer">This is actually working</p></footer>
 </html>
